Update PlayerbotFactory.cpp
code should move to src/RandomPlayerbotMgr.cpp,It is not about initializing arena team members, but rather de-initializing them. The initialization of arena team members should be completed in one go, instead of being done in multiple batches. If it is carried out in multiple batches, it may result in arena teams not being fully staffed, which will increase the queue time for high-score arenas.